Extract Batch 4 read/mutation tool handlers into src/tools/
Moves forget, get, where_is_memory, list, and discover_tags handlers
from src/index.ts into dedicated modules under src/tools/.

Each module exports registerXxxTool(server, ctx) following the
established ctx-based dependency injection pattern.
