+++ src/index.ts
@@ -48,6 +48,11 @@ import { registerSetProjectMemoryPolicyTool, registerGetProjectMemoryPolicyTool 
 import { registerRememberTool } from "./tools/remember.js";
 import { registerRecallTool } from "./tools/recall.js";
 import { registerUpdateTool } from "./tools/update.js";
+import { registerForgetTool } from "./tools/forget.js";
+import { registerGetTool } from "./tools/get.js";
+import { registerWhereIsMemoryTool } from "./tools/where-is-memory.js";
+import { registerListTool } from "./tools/list.js";
+import { registerDiscoverTagsTool } from "./tools/discover-tags.js";
 import { suggestAutoRelationships } from "./auto-relate.js";
 import {
   computeHybridScore,
@@ -617,685 +622,15 @@ registerRecallTool(server, ctx);
 
 registerUpdateTool(server, ctx);
 
-// ── forget ────────────────────────────────────────────────────────────────────
-server.registerTool(
-  "forget",
-  {
-    title: "Forget",
-    description:
-      "Delete an existing memory by id and clean up dangling relationships.\n\n" +
-      "Use this when:\n" +
-      "- A memory should be removed entirely\n" +
-      "- A note is obsolete and should not remain searchable\n\n" +
-      "Do not use this when:\n" +
-      "- The note should stay but move to another vault; use `move_memory`\n" +
-      "- The note should be replaced or merged; use `consolidate`\n\n" +
-      "Returns:\n" +
-      "- Deleted memory id/title and relationship cleanup details\n\n" +
-      "Side effects: deletes note files, cleans relationship references, git commits, and may push.\n\n" +
-      "Typical next step:\n" +
-      "- Use `recall` or `list` to confirm the remaining memory set is clean.",
-    annotations: {
-      readOnlyHint: false,
-      destructiveHint: true,
-      idempotentHint: true,
-      openWorldHint: false,
-    },
-    inputSchema: z.object({
-      id: NoteIdSchema.describe("Exact memory id. Use an id returned by `recall`, `list`, `recent_memories`, or `where_is`."),
-      cwd: projectParam,
-      allowProtectedBranch: z
-        .boolean()
-        .optional()
-        .describe(
-          "One-time override for protected branch checks. " +
-          "When true, forget can commit on a protected branch without changing project policy."
-        ),
-    }),
-    outputSchema: ForgetResultSchema,
-  },
-  async ({ id, cwd, allowProtectedBranch = false }) => {
-    await ensureBranchSynced(cwd);
+registerForgetTool(server, ctx);
 
-    const found = await vaultManager.findNote(id, cwd);
-    if (!found) {
-      return { content: [{ type: "text", text: `No memory found with id '${id}'` }], isError: true };
-    }
+registerGetTool(server, ctx);
 
-    const { note, vault: noteVault } = found;
-    const touchesProjectVault = noteVault.isProject || await wouldRelationshipCleanupTouchProjectVault([id]);
-    if (touchesProjectVault) {
-      const resolvedProject = await resolveProject(cwd);
-      const projectLabel = resolvedProject
-        ? `${resolvedProject.name} (${resolvedProject.id})`
-        : `${note.projectName ?? "project"} (${note.project ?? "unknown"})`;
-      const policy = note.project ? await configStore.getProjectPolicy(note.project) : undefined;
-      const protectedBranchCheck = await shouldBlockProtectedBranchCommit({
-        cwd,
-        writeScope: "project",
-        automaticCommit: true,
-        projectLabel,
-        policy,
-        allowProtectedBranch,
-        toolName: "forget",
-      });
-      if (protectedBranchCheck.blocked) {
-        return {
-          content: [{ type: "text", text: protectedBranchCheck.message ?? "Protected branch policy blocked this commit." }],
-          isError: true,
-        };
-      }
-    }
+registerWhereIsMemoryTool(server, ctx);
 
-    await noteVault.storage.deleteNote(memoryId(id));
+registerListTool(server, ctx);
 
-    // Clean up dangling references grouped by vault so we make one commit per vault
-    const vaultChanges = await removeRelationshipsToNoteIds([id]);
-
-    // Always include the deleted note's path (git add on a deleted file stages the removal)
-    addVaultChange(vaultChanges, noteVault, vaultManager.noteRelPath(noteVault, id));
-
-    let retry: MutationRetryContract | undefined;
-    for (const [v, files] of vaultChanges) {
-      const isPrimaryVault = v === noteVault;
-      const summary = isPrimaryVault ? `Deleted note and cleaned up ${files.length - 1} reference(s)` : "Cleaned up dangling reference";
-      const commitBody = formatCommitBody({
-        summary,
-        noteId: id,
-        noteTitle: note.title,
-        projectName: note.projectName,
-      });
-      const commitMessage = `forget: ${note.title}`;
-      const commitStatus = await v.git.commitWithStatus(commitMessage, files, commitBody);
-      if (!retry) {
-        retry = buildMutationRetryContract({
-          commit: commitStatus,
-          commitMessage,
-          commitBody,
-          files,
-          cwd,
-          vault: v,
-          mutationApplied: true,
-        });
-      }
-      if (commitStatus.status === "committed") {
-        await pushAfterMutation(v);
-      }
-    }
-
-    const structuredContent: ForgetResult = {
-      action: "forgotten",
-      id,
-      title: note.title,
-      project: noteProjectRef(note),
-      relationshipsCleaned: vaultChanges.size > 0 ? Array.from(vaultChanges.values()).reduce((sum, files) => sum + files.length - 1, 0) : 0,
-      vaultsModified: Array.from(vaultChanges.keys()).map(v => storageLabel(v)),
-      retry,
-    };
-    
-    const retrySummary = formatRetrySummary(retry);
-    invalidateActiveProjectCache();
-    return {
-      content: [{
-        type: "text",
-        text: `Forgotten '${id}' (${note.title})${retrySummary ? `\n${retrySummary}` : ""}`,
-      }],
-      structuredContent,
-    };
-  }
-);
-
-// ── get ───────────────────────────────────────────────────────────────────────
-server.registerTool(
-  "get",
-  {
-    title: "Get Memory",
-    description:
-      "Use after `recall`, `list`, or `recent_memories` when you need the full note content.\n\n" +
-      "Fetch one or more memories by exact id.\n\n" +
-      "Use this when:\n" +
-      "- You already know the memory id and need the full note content\n" +
-      "- A previous tool returned ids that you now want to inspect exactly\n\n" +
-      "Do not use this when:\n" +
-      "- You are still searching by topic; use `recall`\n" +
-      "- You want to browse many notes; use `list`\n\n" +
-      "Returns:\n" +
-      "- Full note content and metadata for the requested ids, including storage label\n" +
-      "- Bounded 1-hop relationship previews when `includeRelationships` is true (max 3 shown)\n\n" +
-      "Read-only.\n\n" +
-      "Typical next step:\n" +
-      "- Use `update`, `forget`, `move_memory`, or `relate` after inspection.",
-    annotations: {
-      readOnlyHint: true,
-      destructiveHint: false,
-      idempotentHint: true,
-      openWorldHint: false,
-    },
-    inputSchema: z.object({
-      ids: z.array(NoteIdSchema).min(1).describe("One or more memory ids to fetch"),
-      cwd: projectParam,
-      includeRelationships: z.boolean().optional().default(false).describe("Include bounded direct relationship previews (1-hop expansion, max 3 shown)"),
-    }),
-    outputSchema: GetResultSchema,
-  },
-  async ({ ids, cwd, includeRelationships }) => {
-    const t0Get = performance.now();
-    await ensureBranchSynced(cwd);
-
-    const project = await resolveProject(cwd);
-    const found: GetResult["notes"] = [];
-    const notFound: string[] = [];
-
-    for (const id of ids) {
-      // Check session cache before hitting storage
-      let result: { note: Note; vault: Vault } | null = null;
-      if (project) {
-        for (const vault of vaultManager.allKnownVaults()) {
-          const cached = getSessionCachedNote(project.id, vault.storage.vaultPath, id);
-          if (cached !== undefined) {
-            result = { note: cached, vault };
-            break;
-          }
-        }
-      }
-      if (!result) {
-        result = await vaultManager.findNote(id, cwd);
-      }
-      if (!result) {
-        notFound.push(id);
-        continue;
-      }
-      const { note, vault } = result;
-      
-      let relationships: RelationshipPreview | undefined;
-      if (includeRelationships) {
-        relationships = await getRelationshipPreview(
-          note,
-          vaultManager.allKnownVaults(),
-          { activeProjectId: project?.id, limit: 3 }
-        );
-      }
-      
-      found.push({
-        id: note.id,
-        title: note.title,
-        content: note.content,
-        project: noteProjectRef(note),
-        tags: note.tags,
-        lifecycle: note.lifecycle,
-        role: note.role,
-        alwaysLoad: note.alwaysLoad,
-        relatedTo: note.relatedTo,
-        createdAt: note.createdAt,
-        updatedAt: note.updatedAt,
-        vault: storageLabel(vault),
-        relationships,
-      });
-
-      if (project) {
-        setSessionCachedNote(project.id, vault.storage.vaultPath, note);
-        recordSessionNoteAccess(project.id, vault.storage.vaultPath, note.id, "get");
-      }
-    }
-
-    const lines: string[] = [];
-    for (const note of found) {
-      lines.push(`## ${note.title} (${note.id})`);
-      lines.push(`project: ${note.project?.name ?? "global"} | stored: ${note.vault} | lifecycle: ${note.lifecycle}${note.role ? ` | role: ${note.role}` : ""}`);
-      if (note.tags.length > 0) lines.push(`tags: ${note.tags.join(", ")}`);
-      lines.push("");
-      lines.push(note.content);
-      if (note.relationships) {
-        lines.push("");
-        lines.push(formatRelationshipPreview(note.relationships));
-      }
-      lines.push("");
-    }
-    if (notFound.length > 0) {
-      lines.push(`Not found: ${notFound.join(", ")}`);
-    }
-
-    const structuredContent: GetResult = {
-      action: "got",
-      count: found.length,
-      notes: found,
-      notFound,
-    };
-
-    console.error(`[get:timing] ${(performance.now() - t0Get).toFixed(1)}ms`);
-    return { content: [{ type: "text", text: lines.join("\n").trim() }], structuredContent };
-  }
-);
-
-// ── where_is_memory ───────────────────────────────────────────────────────────
-server.registerTool(
-  "where_is_memory",
-  {
-    title: "Where Is Memory",
-    description:
-      "Locate which storage label and project association a memory currently has.\n\n" +
-      "Use this when:\n" +
-      "- You know the id and want to see where the memory lives\n" +
-      "- You are deciding whether a note should be moved between project storage locations\n\n" +
-      "Do not use this when:\n" +
-      "- You need the full note content; use `get`\n" +
-      "- You want to search for a note by topic; use `recall`\n\n" +
-      "Returns:\n" +
-      "- Title, project association, storage label, updated time, and relationship count\n\n" +
-      "Read-only.\n\n" +
-      "Typical next step:\n" +
-      "- Use `move_memory` if the storage location is wrong, or `get` for full inspection.",
-    annotations: {
-      readOnlyHint: true,
-      destructiveHint: false,
-      idempotentHint: true,
-      openWorldHint: false,
-    },
-    inputSchema: z.object({
-      id: NoteIdSchema.describe("Exact memory id. Use an id returned by `recall`, `list`, `recent_memories`, or `where_is`."),
-      cwd: projectParam,
-    }),
-    outputSchema: WhereIsResultSchema,
-  },
-  async ({ id, cwd }) => {
-    await ensureBranchSynced(cwd);
-
-    const found = await vaultManager.findNote(id, cwd);
-    if (!found) {
-      return { content: [{ type: "text", text: `No memory found with id '${id}'` }], isError: true };
-    }
-
-    const { note, vault } = found;
-    const vaultLabel = storageLabel(vault);
-    const projectDisplay = note.projectName && note.project
-      ? `${note.projectName} (${note.project})`
-      : note.projectName ?? note.project ?? "global";
-    const relatedCount = note.relatedTo?.length ?? 0;
-
-    const structuredContent: WhereIsResult = {
-      action: "located",
-      id: note.id,
-      title: note.title,
-      project: noteProjectRef(note),
-      vault: vaultLabel,
-      updatedAt: note.updatedAt,
-      relatedCount,
-    };
-
-    return {
-      content: [{
-        type: "text",
-        text: `'${note.title}' (${id})\nproject: ${projectDisplay} | stored: ${vaultLabel} | updated: ${note.updatedAt} | related: ${relatedCount}`,
-      }],
-      structuredContent,
-    };
-  }
-);
-
-// ── list ──────────────────────────────────────────────────────────────────────
-server.registerTool(
-  "list",
-  {
-    title: "List Memories",
-    description:
-      "List stored memories with filtering by scope, storage label, and tags.\n\n" +
-      "Use this when:\n" +
-      "- You want to browse what exists for a project or globally\n" +
-      "- You want a deterministic filtered list rather than a semantic search\n" +
-      "- You are checking inventory before creating, updating, or consolidating notes\n\n" +
-      "Do not use this when:\n" +
-      "- You want topic-based semantic search; use `recall`\n" +
-      "- You already know the exact id; use `get`\n\n" +
-      "Returns:\n" +
-      "- Matching memories with ids, titles, scope/storage context, and metadata\n\n" +
-      "Read-only.\n\n" +
-      "Typical next step:\n" +
-      "- Use `get` for exact inspection or `update` / `consolidate` for cleanup.",
-    annotations: {
-      readOnlyHint: true,
-      destructiveHint: false,
-      idempotentHint: true,
-      openWorldHint: false,
-    },
-    inputSchema: z.object({
-      cwd: projectParam,
-      scope: z
-        .enum(["project", "global", "all"])
-        .optional()
-        .default("all")
-        .describe(
-          "'project' = only this project's memories (project-scoped storage); " +
-          "'global' = only unscoped memories (main/global storage); " +
-          "'all' = everything visible from this context (default)"
-        ),
-      storedIn: z
-        .enum(["project-vault", "main-vault", "any"])
-        .optional()
-        .default("any")
-        .describe(
-          "Storage-label filter. Use `main-vault` for main/global storage. " +
-          "Use `project-vault` as the broad filter for any project vault, including sub-vaults. " +
-          "Results may still return a more specific label such as `sub-vault:.mnemonic-lib`."
-        ),
-      tags: z.array(z.string()).optional().describe("Filter to notes matching all of these tags."),
-      includeRelations: z.boolean().optional().default(false).describe("Include related memory ids and relationship types"),
-      includePreview: z.boolean().optional().default(false).describe("Include a short content preview for each note"),
-      includeStorage: z.boolean().optional().default(false).describe("Show which vault each note is stored in"),
-      includeUpdated: z.boolean().optional().default(false).describe("Include last-updated timestamp for each note"),
-    }),
-    outputSchema: ListResultSchema,
-  },
-  async ({ cwd, scope, storedIn, tags, includeRelations, includePreview, includeStorage, includeUpdated }) => {
-    await ensureBranchSynced(cwd);
-
-    const { project, entries } = await collectVisibleNotes(cwd, scope, tags, storedIn);
-
-    if (entries.length === 0) {
-      const structuredContent: ListResult = { action: "listed", count: 0, scope: scope || "all", storedIn: storedIn || "any", project: project ? { id: project.id, name: project.name } : undefined, notes: [] };
-      return { content: [{ type: "text", text: "No memories found." }], structuredContent };
-    }
-
-    const lines = entries.map((entry) => formatListEntry(entry, {
-      includeRelations,
-      includePreview,
-      includeStorage,
-      includeUpdated,
-    }));
-
-    const header = project && scope !== "global"
-      ? `${entries.length} memories (project: ${project.name}, scope: ${scope}, storedIn: ${storedIn}):`
-      : `${entries.length} memories (scope: ${scope}, storedIn: ${storedIn}):`;
-
-    const textContent = `${header}\n\n${lines.join("\n")}`;
-    
-    const structuredNotes: Array<{
-      id: string;
-      title: string;
-      project?: ProjectRef;
-      tags: string[];
-      lifecycle: NoteLifecycle;
-      role?: Note["role"];
-      vault: string;
-      updatedAt: string;
-      hasRelated?: boolean;
-    }> = entries.map(({ note, vault }) => ({
-      id: note.id,
-      title: note.title,
-      project: noteProjectRef(note),
-      tags: note.tags,
-      lifecycle: note.lifecycle,
-      role: note.role,
-      vault: storageLabel(vault),
-      updatedAt: note.updatedAt,
-      hasRelated: note.relatedTo && note.relatedTo.length > 0,
-    }));
-    
-    const structuredContent: ListResult = {
-      action: "listed",
-      count: entries.length,
-      scope: scope || "all",
-      storedIn: storedIn || "any",
-      project: project ? { id: project.id, name: project.name } : undefined,
-      notes: structuredNotes,
-      options: {
-        includeRelations,
-        includePreview,
-        includeStorage,
-        includeUpdated,
-      },
-    };
-
-    return { content: [{ type: "text", text: textContent }], structuredContent };
-  }
-);
-
-// ── discover_tags ───────────────────────────────────────────────────────────
-server.registerTool(
-  "discover_tags",
-  {
-    title: "Discover Tags",
-    description:
-      "Suggest canonical tags for a specific note before `remember` when tag choice is ambiguous.\n\n" +
-      "Use this when:\n" +
-      "- You have a note title, content, or query and want compact tag suggestions\n" +
-      "- You want canonical project terminology without exposing lots of unrelated tags\n" +
-      "- You want to demote temporary-only tags unless they fit the note\n\n" +
-      "Do not use this when:\n" +
-      "- You need to browse notes by tag; use `list` with `tags` filter instead\n" +
-      "- You already know the exact tags you want to use\n" +
-      "- You want broad inventory output but are not explicitly requesting `mode: \"browse\"`\n\n" +
-      "Returns:\n" +
-      "- Default: bounded `recommendedTags` ranked by note relevance first and usage count second\n" +
-      "- Each suggestion includes canonicality and lifecycle signals plus one compact example\n" +
-      "- Optional `mode: \"browse\"` returns broader inventory output\n\n" +
-      "Typical next step:\n" +
-      "- Reuse suggested canonical tags when they fit, or create a new tag only when genuinely novel.\n\n" +
-      "Performance: O(n) where n = total notes scanned. Expect 100-200ms for 500 notes.\n\n" +
-      "Read-only.",
-    annotations: {
-      readOnlyHint: true,
-      destructiveHint: false,
-      idempotentHint: true,
-      openWorldHint: false,
-    },
-    inputSchema: z.object({
-      cwd: projectParam,
-      mode: z.enum(["suggest", "browse"]).optional().default("suggest"),
-      title: z.string().optional(),
-      content: z.string().optional(),
-      query: z.string().optional(),
-      candidateTags: z.array(z.string()).optional(),
-      lifecycle: z.enum(NOTE_LIFECYCLES).optional(),
-      scope: z
-        .enum(["project", "global", "all"])
-        .optional()
-        .default("all")
-        .describe(
-          "'project' = only this project's memories; " +
-          "'global' = only unscoped memories; " +
-          "'all' = everything visible (default)"
-        ),
-      storedIn: z
-        .enum(["project-vault", "main-vault", "any"])
-        .optional()
-        .default("any")
-        .describe("Filter by vault storage label like list tool."),
-      limit: z.number().int().min(1).max(50).optional(),
-    }),
-    outputSchema: DiscoverTagsResultSchema,
-  },
-  async ({ cwd, mode, title, content, query, candidateTags, lifecycle, scope, storedIn, limit }) => {
-    await ensureBranchSynced(cwd);
-    const startTime = Date.now();
-
-    const { project, entries } = await collectVisibleNotes(cwd, scope, undefined, storedIn);
-
-    const tagStats = new Map<string, DiscoverTagStat>();
-    const candidateTagSet = new Set((candidateTags || []).map(tag => tag.toLowerCase()));
-    const contextValues = [title, content, query, ...(candidateTags || [])];
-    const contextTokens = new Set(tokenizeTagDiscoveryText(contextValues.filter(Boolean).join(" ")));
-    const isTemporaryTarget = lifecycle === "temporary";
-    const effectiveLimit = limit ?? (mode === "browse" ? 20 : 10);
-
-    for (const { note } of entries) {
-      const noteTokens = new Set(tokenizeTagDiscoveryText(`${note.title} ${note.content} ${note.tags.join(" ")}`));
-      const contextMatches = contextTokens.size > 0 ? countTokenOverlap(contextTokens, noteTokens) : 0;
-
-      for (const tag of note.tags) {
-        const stats = tagStats.get(tag) || {
-          count: 0,
-          examples: [],
-          lifecycles: new Set<NoteLifecycle>(),
-          contextMatches: 0,
-          exactCandidateMatch: false,
-        };
-        stats.count++;
-        if (stats.examples.length < 3) {
-          stats.examples.push(note.title);
-        }
-        stats.lifecycles.add(note.lifecycle);
-        stats.contextMatches += contextMatches;
-        if (candidateTagSet.has(tag.toLowerCase())) {
-          stats.exactCandidateMatch = true;
-        }
-        tagStats.set(tag, stats);
-      }
-    }
-
-    const rawTags = Array.from(tagStats.entries())
-      .map(([tag, stats]) => {
-        const lifecycleTypes = Array.from(stats.lifecycles) as NoteLifecycle[];
-        const isTemporaryOnly = stats.lifecycles.size === 1 && stats.lifecycles.has("temporary");
-        const tagTokens = tokenizeTagDiscoveryText(tag);
-        const exactContextMatch = stats.exactCandidateMatch || hasExactTagContextMatch(tag, contextValues);
-        const tagTokenOverlap = contextTokens.size > 0
-          ? countTokenOverlap(contextTokens, tagTokens)
-          : 0;
-        const averageContextMatch = stats.count > 0 ? stats.contextMatches / stats.count : 0;
-        const reason = exactContextMatch
-          ? "matches a candidate tag already present in the note context"
-          : tagTokenOverlap > 0 || stats.contextMatches > 0
-            ? "matches the note context and existing project usage"
-            : "high-usage canonical tag from existing project notes";
-
-        return {
-          tag,
-          usageCount: stats.count,
-          examples: stats.examples,
-          example: stats.examples[0],
-          reason,
-          lifecycleTypes,
-          isTemporaryOnly,
-          exactContextMatch,
-          tagTokenOverlap,
-          averageContextMatch,
-          isBroadSingleToken: tagTokens.length === 1,
-          isHighFrequency: stats.count >= 4,
-          specificityBoost: tagTokens.length > 1 ? 4 : 0,
-        };
-      });
-
-    const hasStrongSpecificCandidate = rawTags.some(tag =>
-      tag.exactContextMatch && (!tag.isBroadSingleToken || tag.usageCount > 1)
-    );
-
-    const tags = rawTags
-      .map((tag) => {
-        const hasWeakDirectMatch = tag.tagTokenOverlap <= 1 && tag.averageContextMatch <= 2;
-        const genericPenalty = hasStrongSpecificCandidate && tag.isBroadSingleToken && tag.isHighFrequency && hasWeakDirectMatch
-          ? 10
-          : 0;
-        const score = hasStrongSpecificCandidate
-          ? (tag.exactContextMatch ? 12 : 0) +
-            (tag.tagTokenOverlap * 5) +
-            tag.specificityBoost +
-            (tag.averageContextMatch * 3) +
-            (tag.usageCount * 0.1) -
-            genericPenalty -
-            (!isTemporaryTarget && tag.isTemporaryOnly ? 2 : 0)
-          : (tag.exactContextMatch ? 6 : 0) +
-            (tag.tagTokenOverlap * 3) +
-            (tag.averageContextMatch * 2) +
-            (tag.usageCount * 1.5) +
-            (tag.isTemporaryOnly ? -3 : 1) -
-            (!isTemporaryTarget && tag.isTemporaryOnly ? 2 : 0);
-
-        return {
-          ...tag,
-          score,
-        };
-      })
-      .sort((a, b) => b.score - a.score || b.usageCount - a.usageCount || a.tag.localeCompare(b.tag));
-
-    const durationMs = Date.now() - startTime;
-    const vaultsSearched = new Set(entries.map(e => storageLabel(e.vault))).size;
-
-    const lines: string[] = [];
-    if (mode === "browse") {
-      if (project && scope !== "global") {
-        lines.push(`Tags for ${project.name} (scope: ${scope}):`);
-      } else {
-        lines.push(`Tags (scope: ${scope}):`);
-      }
-      lines.push("");
-      lines.push(`Total: ${tags.length} unique tags across ${entries.length} notes (${durationMs}ms)`);
-      lines.push("");
-      lines.push("Tags sorted by usage:");
-      for (const t of tags.slice(0, effectiveLimit)) {
-        const lifecycleMark = t.isTemporaryOnly ? " [temp-only]" : "";
-        lines.push(`  ${t.tag} (${t.usageCount})${lifecycleMark}`);
-        if (t.examples.length > 0) {
-          lines.push(`    Example: "${t.examples[0]}"`);
-        }
-      }
-      if (tags.length > effectiveLimit) {
-        lines.push(`  ... and ${tags.length - effectiveLimit} more`);
-      }
-    } else {
-      const suggestedTags = tags.slice(0, effectiveLimit);
-      if (project && scope !== "global") {
-        lines.push(`Suggested tags for ${project.name} (scope: ${scope}):`);
-      } else {
-        lines.push(`Suggested tags (scope: ${scope}):`);
-      }
-      lines.push("");
-      lines.push(`Considered ${tags.length} unique tags across ${entries.length} notes (${durationMs}ms)`);
-      lines.push("");
-      if (contextTokens.size === 0) {
-        lines.push("No note context provided; showing the most canonical existing tags.");
-        lines.push("");
-      }
-      lines.push("Recommended tags:");
-      for (const t of suggestedTags) {
-        const lifecycleMark = t.isTemporaryOnly ? " [temp-only]" : "";
-        lines.push(`  ${t.tag} (${t.usageCount})${lifecycleMark}`);
-        if (t.example) {
-          lines.push(`    Example: "${t.example}"`);
-        }
-        lines.push(`    Why: ${t.reason}`);
-      }
-    }
-
-    const structuredContent: DiscoverTagsResult = mode === "browse"
-      ? {
-          action: "tags_discovered",
-          project: project ? { id: project.id, name: project.name } : undefined,
-          mode,
-          scope: scope || "all",
-          tags: tags.slice(0, effectiveLimit).map(tag => ({
-            tag: tag.tag,
-            usageCount: tag.usageCount,
-            examples: tag.examples,
-            lifecycleTypes: tag.lifecycleTypes,
-            isTemporaryOnly: tag.isTemporaryOnly,
-          })),
-          totalTags: tags.length,
-          totalNotes: entries.length,
-          vaultsSearched,
-          durationMs,
-        }
-      : {
-          action: "tags_discovered",
-          project: project ? { id: project.id, name: project.name } : undefined,
-          mode,
-          scope: scope || "all",
-          recommendedTags: tags.slice(0, effectiveLimit).map(tag => ({
-            tag: tag.tag,
-            usageCount: tag.usageCount,
-            example: tag.example,
-            reason: tag.reason,
-            lifecycleTypes: tag.lifecycleTypes,
-            isTemporaryOnly: tag.isTemporaryOnly,
-          })),
-          totalTags: tags.length,
-          totalNotes: entries.length,
-          vaultsSearched,
-          durationMs,
-        };
-
-    return { content: [{ type: "text", text: lines.join("\n") }], structuredContent };
-  }
-);
+registerDiscoverTagsTool(server, ctx);
 
 // ── recent_memories ───────────────────────────────────────────────────────────
 server.registerTool(
